poc-2/requirements.txt
faiss-cpu
sentence_transformers            # encoding model
pymupdf                          # PDF reading / form widgets
pandas                           # CSV helpers (index build)